"Eoulsan version 2.2"

+++ src/main/java/fr/ens/biologie/genomique/eoulsan/bio/readsmappers/MapperProcess.java
@@ -25,6 +25,7 @@
 package fr.ens.biologie.genomique.eoulsan.bio.readsmappers;
 
 import static fr.ens.biologie.genomique.eoulsan.EoulsanLogger.getLogger;
+import static java.util.Objects.requireNonNull;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -60,20 +61,23 @@ public abstract class MapperProcess {
   private final MapperExecutor executor;
   private final boolean pairedEnd;
 
-  private List<Result> processResults = new ArrayList<>();
+  private final List<Result> processResults = new ArrayList<>();
 
   private InputStream stdout;
 
   private final File pipeFile1;
   private final File pipeFile2;
+  private final File stdErrFile;
 
   private final FastqWriter writer1;
   private final FastqWriter writer2;
 
+  private final File temporaryDirectory;
+
   private ReporterIncrementer incrementer;
   private String counterGroup;
 
-  private List<File> filesToRemove = new ArrayList<>();
+  private final List<File> filesToRemove = new ArrayList<>();
 
   //
   // Inner classes
@@ -717,6 +721,20 @@ public abstract class MapperProcess {
   }
 
   /**
+   * Start the process.
+   * @throws IOException if an error occurs while starting the process
+   */
+  void startProcess() throws IOException {
+
+    try {
+      // Start mapper instance
+      startProcess(this.temporaryDirectory);
+    } catch (InterruptedException e) {
+      throw new IOException(e);
+    }
+  }
+
+  /**
    * Start the process(es) of the mapper.
    * @param tmpDirectory temporary directory
    * @throws IOException if an error occurs while starting the process(es)
@@ -735,8 +753,9 @@ public abstract class MapperProcess {
 
       final boolean last = i == cmds.size() - 1;
 
-      final Result result = this.executor.execute(cmds.get(i),
-          executionDirectory, last, false, this.pipeFile1, this.pipeFile2);
+      final Result result =
+          this.executor.execute(cmds.get(i), executionDirectory, last,
+              this.stdErrFile, false, this.pipeFile1, this.pipeFile2);
 
       this.processResults.add(result);
 
@@ -780,7 +799,7 @@ public abstract class MapperProcess {
    */
   private void removeFile(final File f) {
 
-    if (f.exists()) {
+    if (f != null && f.exists()) {
 
       if (!f.delete()) {
         getLogger().warning("Cannot remove temporary file: " + f);
@@ -830,42 +849,112 @@ public abstract class MapperProcess {
 
   /**
    * Constructor.
-   * @param mapper mapper to use
+   * @param mapperName mapper name
+   * @param executor executor
+   * @param temporaryDirectory temporary directory
    * @param pairedEnd paired-end mode
-   * @throws IOException if en error occurs
+   * @throws IOException if an error occurs
    */
-  protected MapperProcess(final AbstractSequenceReadsMapper mapper,
+  protected MapperProcess(final String mapperName, MapperExecutor executor,
+      final File temporaryDirectory, final File stdErrFile,
       final boolean pairedEnd) throws IOException {
 
-    this(mapper, pairedEnd, false);
+    this(mapperName, executor, temporaryDirectory, stdErrFile, pairedEnd,
+        false);
   }
 
   /**
    * Constructor.
-   * @param mapper mapper to use
+   * @param mapperName mapper name
+   * @param executor executor
+   * @param temporaryDirectory temporary directory
+   * @param pairedEnd paired-end mode
+   * @param inputFile1 first file to map
+   * @param inputFile2 second file to map
+   * @throws IOException if an error occurs
+   */
+  protected MapperProcess(final String mapperName, MapperExecutor executor,
+      final File temporaryDirectory, final File stdErrFile,
+      final boolean pairedEnd, final File inputFile1, final File inputFile2)
+      throws IOException {
+
+    this(mapperName, executor, temporaryDirectory, stdErrFile, pairedEnd, false,
+        inputFile1, inputFile2);
+  }
+
+  /**
+   * Constructor.
+   * @param mapperName mapper name
+   * @param executor executor
+   * @param temporaryDirectory temporary directory
+   * @param pairedEnd paired-end mode
+   * @param inputFile first file to map
+   * @throws IOException if an error occurs
+   */
+  protected MapperProcess(final String mapperName, MapperExecutor executor,
+      final File temporaryDirectory, final File stdErrFile,
+      final boolean pairedEnd, final File inputFile) throws IOException {
+
+    this(mapperName, executor, temporaryDirectory, stdErrFile, pairedEnd, false,
+        inputFile, null);
+  }
+
+  /**
+   * Constructor.
+   * @param mapperName mapper name
+   * @param executor executor
+   * @param temporaryDirectory temporary directory
    * @param pairedEnd paired-end mode
    * @throws IOException if en error occurs
    */
-  protected MapperProcess(final AbstractSequenceReadsMapper mapper,
+  protected MapperProcess(final String mapperName, MapperExecutor executor,
+      final File temporaryDirectory, final File stdErrFile,
       final boolean pairedEnd, final boolean threadForRead1)
       throws IOException {
 
-    if (mapper == null) {
-      throw new NullPointerException("The mapper is null");
-    }
+    this(mapperName, executor, temporaryDirectory, stdErrFile, pairedEnd,
+        threadForRead1, null, null);
+  }
 
-    try {
-      this.mapperName = mapper.getMapperName();
-      this.uuid = UUID.randomUUID().toString();
+  /**
+   * Constructor.
+   * @param mapperName mapper name
+   * @param executor executor
+   * @param temporaryDirectory temporary directory
+   * @param pairedEnd paired-end mode
+   * @param inputFile1 first file to map
+   * @param inputFile2 second file to map
+   * @throws IOException if en error occurs
+   */
+  protected MapperProcess(final String mapperName, MapperExecutor executor,
+      final File temporaryDirectory, final File stdErrFile,
+      final boolean pairedEnd, final boolean threadForRead1,
+      final File inputFile1, final File inputFile2) throws IOException {
 
-      this.executor = mapper.getExecutor();
-      this.pairedEnd = pairedEnd;
+    requireNonNull(mapperName, "mapperName argument cannot be null");
+    requireNonNull(executor, "executor argument cannot be null");
+    requireNonNull(temporaryDirectory,
+        "temporaryDirectory argument cannot be null");
 
-      // Define temporary files
-      final File tmpDir = mapper.getTempDirectory();
+    this.mapperName = mapperName;
+    this.uuid = UUID.randomUUID().toString();
 
-      this.pipeFile1 = new File(tmpDir, "mapper-inputfile1-" + uuid + ".fq");
-      this.pipeFile2 = new File(tmpDir, "mapper-inputfile2-" + uuid + ".fq");
+    this.executor = executor;
+    this.pairedEnd = pairedEnd;
+
+    this.temporaryDirectory = temporaryDirectory;
+
+    this.pipeFile1 = inputFile1 != null
+        ? inputFile1 : new File(this.temporaryDirectory,
+            "mapper-inputfile1-" + uuid + ".fq");
+    this.pipeFile2 = inputFile2 != null
+        ? inputFile2 : new File(this.temporaryDirectory,
+            "mapper-inputfile2-" + uuid + ".fq");
+
+    this.stdErrFile = stdErrFile;
+
+    // If in entry mode
+    if (inputFile1 == null) {
 
       this.writer1 = threadForRead1
           ? new FastqWriterThread(this.pipeFile1, "FastqWriterThread fastq1")
@@ -875,14 +964,12 @@ public abstract class MapperProcess {
           : null;
 
       addFilesToRemove(this.pipeFile1, this.pipeFile2);
-
-      additionalInit();
-
-      // Start mapper instance
-      startProcess(tmpDir);
-
-    } catch (InterruptedException e) {
-      throw new IOException(e);
+    } else {
+      this.writer1 = null;
+      this.writer2 = null;
     }
+
+    // Launch addition initialization
+    additionalInit();
   }
 }
